Fetch logs in chunks (#29)
When querying an RPC node for huge range of events it can return an
error when the response size gets exceeded.
This error prevented the fee withdrawal to work correctly in a week
where a lot of trades happened.

This PR works around this limitation by requesting logs in chunks
instead of the whole range at once.

Mentioned error:
```
Error: Log response size exceeded. You can make eth_getLogs requests with up to a 2K block range and no limit on the response size, or you can request any block range with a cap of 10K logs in the response. Based on your parameters and the response size limit, this block range should work: [0x13b3ac6, 0x13b6ec9]
```

---------

Co-authored-by: Felix Leupold <felixleupold90@gmail.com>