feat: automatically wrap native tokens (#40)
This PR is a WIP

It aims to add automatic extraction of the native balance too. 

The new logic is:
- If there's more than `minOut` of native tokens they are wrapped
- If there's more than `minOut` of wrapped native tokens, they are
transferred to the receiver. Note that because the previous wrap
interaction is executed first, the transfer includes the wrapped native
token, and the wrapped native token it already had.

This PR removes the flexibility of setting an arbitrary `toToken` and
instead renames the param to `wrappedNativeToken`. This allows to do
simplifications to implement this feature.

## Test
The contracts have been deployed to Sepolia. 

This is an example of a `drip` with no wrapping (there was not enough
ETH).

https://sepolia.etherscan.io/tx/0xd65fbfa81baf46a85ca6962f569f9c8cea50e275dd6382fc4261085c9e9f7cd1

I gifted 1 ETH to the contract, and re-run `drip` and that extracted it:

https://sepolia.etherscan.io/tx/0x1ea6ea8ab0ddc773af3dc825205f2503bdc30000bed1a0150fb7e247cef26be6


I'm Sepolia rich now 🤑💰 . Let me know if you need ETH!
<img width="956" alt="image"
src="https://github.com/user-attachments/assets/cbeefd7e-229f-4738-9d58-ae52d1e011b8"
/>


## Context
Before this PR, the module handle the `toToken` in the following way
- If there's balance for the token above the `minOut` threshold, a new
interaction is added to the settlement to transfer the `toToken` to the
receiver (no order involved)
- There was no way to withdraw native currency

In practice, this contract is used with `toToken` set to the native
wrapped token (i.e. WETH in mainnet).

It makes sense that the `toToken` is the native wrapped, because most of
the costs are in the native currency, so it makes sense that the fee is
converted to this token.

When trying to give some way to extract the native token, I saw it
needed to add some complexity. Everything gets simplified if we:
- Remove the `toToken` and add a `wrappedNativeToken` instead. The
`buyToken` is always `wrappedNativeToken`
- If there's native balance above the threshold, we create an
interaction to wrap it
- This way, the transfer interaction can transfer the wrapped amount,
and the `toToken` balance (because now is `wrappedNativeToken` and not
an arbitrary `toToken`)

Given that in practice this module is always instantiated with
`toToken=wrappedNativeToken` , I thought it would be a good compromise
to:
- Make the `drip` call cheaper
- Make the code simpler

## Bonus
I tried to do some small improvements. 

## Script improvements
In the script. Among others
- Show you the calldata and rest of the params before you place the
transaction (this allows you to verify it first). A follow up PR could
add a tenderly link, but I didn't want to deviate so much.
- Adds an optional `-c --confirm-drip` param to ask for confirmation
before sending the transaction. Its `false` by default for making it
backward compatible since I don't know if there's some automatization.
- Now the script asserts that the RPC is in the network matching the
`drip` parameters (avoiding incoherences)
- Added SEPOLIA support (not supported before)
- I broke down some big function into smaller ones
- Improves the messages in general to make them easier to read (i.e.
write Bignumbers directly instead of the string representation)
- Adapt to the latest `app-data` npm package

## Readme improvements
I added some improvements on things I had to do when deploying and I
think other deployers might appreciate:
- Encourage adding `--verify` flag and etherscan API key
- Instructions on how to enable the module in the Safe

## How to review this PR
One good way is to verify the transactions above and read the new unit
tests. The old ones still work, because this PR behaves in the same way
as the previous contract, but with the addition of some automatic
wrapping of Ether.

This PR renames the deployment contract, so there's some files being
moved (in hindsight I should have done that in another PR, sorry).

Basically, it uses a ver for the deployment, similarly to the naming
convention in `flash-loans`:
`broadcast/COWFeeModule.s.sol → broadcast/DeployCOWFeeModule.s.sol`

---------

Co-authored-by: Federico Giacon <58218759+fedgiac@users.noreply.github.com>