chore: deploy mainnet (#50)
Deploys a new module contract with new parameters and the latest feature
to wrap ETH automatically:
~-
https://polygonscan.com/address/0xaF717efCFeC02A65101E3FFB3C07234B8705CAfC~
-
https://etherscan.io/address/0x8429bf5bad17925c364c8c30a3c2a1e0d76602d2#code

## Test
Verify the parameters in the contract (old vs new)

If you use https://github.com/cowdao-grants/cow-fee/pull/49, you will
have to modify
https://github.com/cowdao-grants/cow-fee/blob/deployment-verification-script/compare-deployments.ts#L91
to catch the rejected promise and return `null`. This is because the new
version has a breaking change (rename of `toToken` to
`wrappedNativeToken`.


![image](https://github.com/user-attachments/assets/bae0f452-a2ba-496c-8a6d-df61c83eb87a)