feat: ability to specify fee per gas and priority fee per gas
polygon has been having issues with fee pricing. this seems to be due to an issue with polygon's fee estiamtion mechanism, causing ethers to return incorrect data.

the errors are usually along the lines of:

```
transaction gas price below minimum: gas tip cap 1500000000, minimum needed 25000000000
```

the easiest and most reliable way to resolve this issue is to allow specifying maxFeePerGas and maxPriorityFeePerGas parameters manually.

technically only maxPriorityFeePerGas is required, but we might as well allow specifying both since they are related and may come in handy in the future