feat: ability to override fee per gas and priority fee per gas (#61)
Allow for `maxFeePerGas` and `maxPriorityFeePerGas` to be selectively
overridden

## Context

polygon has been having issues with fee pricing. this seems to be due to
an issue with polygon's fee estiamtion mechanism, causing ethers to
return incorrect data.

the errors are usually along the lines of:

```
transaction gas price below minimum: gas tip cap 1500000000, minimum needed 25000000000
```

the easiest and most reliable way to resolve this issue is to allow
specifying `maxFeePerGas` and `maxPriorityFeePerGas` parameters
manually. Polygon even gives us the exact minimum `maxPriorityFeePerGas`
required, so we can just use that.

technically only `maxPriorityFeePerGas` is required, but we might as
well allow specifying both since they are related and may come in handy
in the future

## Test

This has been tested on Kubernetes using a manual cronjob deployment.
Additional parameters set for polygon are: `--max-priority-fee-per-gas
25` and `--max-fee-per-gas 50`

example of one of the transactions on polygon is here
https://polygonscan.com/tx/0x22b30f39cc69079340173a176e58889d8af52e9e11a4e575705550fc36af12ef
(it sets maxfeepergas to 50 and maxpriorityfeepergas to 25)